+++ b2d-venture/app/(dashboard)/admin/layout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import AdminSideNav from "@/components/shared/AdminSideNav";
+import AdminSideNav from "@/components/shared/AdminDashboard/AdminSideNav";
 
 export default function AdminLayout({ children, params }: { children: React.ReactNode; params: { id: string } }) {
   return (
