+++ b2d-venture/components/shared/AdminDashboard/BusinessRequestCard.tsx
@@ -2,8 +2,8 @@
 import { Card, CardHeader, CardFooter, CardTitle, CardDescription, CardContent } from "@/components/ui/card"
 import Image from "next/image";
 import Link from "next/link";
-import Tag from "../ui/tag";
-import { Button } from "../ui/button";
+import Tag from "../../ui/tag";
+import { Button } from "../../ui/button";
 const BusinessRequestCard = ({className, email, contact, address, name, description, tag, status}) => {
  return (
     <div className ={className}>
